Renaming source files for DashboardClientROS
