+++ ur_robot_driver/include/ur_robot_driver/ros/hardware_interface.h
@@ -49,7 +49,7 @@
 #include <ur_controllers/scaled_joint_command_interface.h>
 
 #include "ur_robot_driver/ur/ur_driver.h"
-#include <ur_robot_driver/ros/dashboard_client.h>
+#include <ur_robot_driver/ros/dashboard_client_ros.h>
 
 namespace ur_driver
 {
